requirements.txt
# Core dependencies - PyTorch nightly for RTX 5080 support
--find-links https://download.pytorch.org/whl/nightly/cu128
torch>=2.2.0.dev0
torchvision>=0.17.0.dev0
torchaudio>=2.2.0.dev0

# GUI framework
PyQt6>=6.5.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# CLIP and quality metrics
transformers>=4.35.0
scikit-image>=0.21.0

# Database
sqlite3  # Built-in with Python

# Utilities
tqdm>=4.66.0
pathlib2>=2.3.7
typing-extensions>=4.8.0

# Parallel processing
joblib>=1.3.0
multiprocessing-logging>=0.3.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 